+++ Citadel/src/utopia/citadel/database/access/single/description/SingleIdDescribedAccess.scala
@@ -3,6 +3,7 @@ package utopia.citadel.database.access.single.description
 import utopia.citadel.database.access.many.description.DescriptionLinksAccess
 import utopia.metropolis.model.cached.LanguageIds
 import utopia.metropolis.model.combined.description.DescribedFactory
+import utopia.metropolis.model.enumeration.DescriptionRoleIdWrapper
 import utopia.metropolis.model.partial.description.DescriptionData
 import utopia.metropolis.model.post.NewDescription
 import utopia.metropolis.model.stored.description.DescriptionLink
@@ -79,6 +80,41 @@ trait SingleIdDescribedAccess[A, +D] extends SingleIntIdModelAccess[A]
 	@deprecated("Please use described instead", "v1.3")
 	def describedInLanguages(languageIds: Seq[Int])(implicit connection: Connection) =
 		described(connection, LanguageIds(languageIds.toVector))
+	
+	/**
+	 * @param roleIds Ids of the targeted description roles
+	 * @param connection Implicit DB Connection
+	 * @param languageIds Accepted language ids
+	 * @return A described copy of this instance, containing 0-1 descriptions per role.
+	 */
+	def describedWithRoleIds(roleIds: Set[Int])(implicit connection: Connection, languageIds: LanguageIds) =
+		pullDescribed { descriptions.withRolesInPreferredLanguages(roleIds) }
+	/**
+	 * @param roles Targeted description roles
+	 * @param connection Implicit DB Connection
+	 * @param languageIds Accepted language ids
+	 * @return A described copy of this instance, containing 0-1 descriptions per role.
+	 */
+	def describedWith(roles: Set[DescriptionRoleIdWrapper])(implicit connection: Connection, languageIds: LanguageIds) =
+		describedWithRoleIds(roles.map { _.id })
+	/**
+	 * @param connection Implicit DB Connection
+	 * @param languageIds Accepted language ids
+	 * @return A described copy of this instance, containing 0-1 descriptions per role.
+	 */
+	def describedWith(firstRole: DescriptionRoleIdWrapper, secondRole: DescriptionRoleIdWrapper,
+	                  moreRoles: DescriptionRoleIdWrapper*)
+	                 (implicit connection: Connection, languageIds: LanguageIds): Option[D] =
+		describedWith(Set(firstRole, secondRole) ++ moreRoles)
+	
+	/**
+	 * @param role Targeted description role
+	 * @param connection Implicit DB Connection
+	 * @param languageIds Accepted language id
+	 * @return A described copy of this instance, containing a description for the specified role, if available
+	 */
+	def withDescription(role: DescriptionRoleIdWrapper)(implicit connection: Connection, languageIds: LanguageIds) =
+		pullDescribed { description.withRole(role).inPreferredLanguage }
 	
 	/**
 	  * Inserts a new description for this item, replacing the existing description