+++ Metropolis/Changes.md
@@ -1,5 +1,10 @@
 # Utopia Metropolis - List of Changes
 
+## v1.3.1 (in development)
+### New Methods
+- **DescribedLanguage**
+  - Added `.descriptionOrCode(...)`
+
 ## v1.3 - 18.10.2021
 This update adds a lot of utility for description handling, unfortunately also breaking the 
 **DescribedFromModelFactory** implementations, although not badly.